+++ CHANGES.md
@@ -5,7 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 See [here](http://keepachangelog.com/) for the change log format.
 
 ## [Unreleased](https://github.com/missingfaktor/akar/compare/v0.1.0...HEAD)
-- Added `if-match` construct
+### Added
+- `if-match` construct
+- `when-match` construct
 
 # v0.1.0 / 2015.07.05